become handbook-first (#300)
* become handbook-first

Introduces the Sourcegraph handbook, which describes how we work. Adds a lot of draft content. The content is subject to change according to the guidelines laid out in the handbook itself.

For now, the handbook is deployed by proxying certain paths (eg `https://about.sourcegraph.com/handbook`) on Netlify to a separate docsite server running the handbook. The deployment will be combined with the rest of the about.sourcegraph.com site cleanly soon.

* remove unused events and hacky-hour doc