confirm new team members are in our GitHub org
I added this to all of the onboarding checklists (which we plan to consolidate into a single checklist) to ensure it's not accidentally forgotten in the consolidation.

The purpose of this is to avoid confusion like at https://sourcegraph.slack.com/archives/CQ44Y7F4G/p1580913355019700 where a new team member had not yet accepted the GitHub invite.