Update gatsby-node.js
@ryan-blunden this updates the conditions in onCreateNode and adds
docs to the file. Hopefully this helps explain the changes I made.

I removed one condition which fixes the blog post case. That condition
was there for an old use case, where we wanted to include the directory
name in the slug. Not actually sure what use case it was for, but I don't
see an applicable case anymore.