Replace Vimeo videos for site content
Note that I've disabled the auto-playing and looping of the main video on the home page as the YouTube iframe embed doesn't allow you to mute audio.

If its vital that this still be autoplayed, I can change the YouTube component to embed using the IFrame API instead of just using an iframe.

Regarding the other Vimeo videos, it's not worth migrating and updating all of the videos in the blog posts due to the small amount of plays they get.