Organize blog posts into folders (#801)
* Categorized blog posts and moved images to website/static/blog

* Update gatsby-node.js

@ryan-blunden this updates the conditions in onCreateNode and adds
docs to the file. Hopefully this helps explain the changes I made.

I removed one condition which fixes the blog post case. That condition
was there for an old use case, where we wanted to include the directory
name in the slug. Not actually sure what use case it was for, but I don't
see an applicable case anymore.

* Moved blogs

* Updated docs for how to add a blog post

* Update README.md

Co-Authored-By: Nick Snyder <nick@sourcegraph.com>

* Moved recent blog posts

* Fixed broken links to Strange Loop posts

Co-authored-by: Farhan Attamimi <farhan@sourcegraph.com>
Co-authored-by: Nick Snyder <nick@sourcegraph.com>