update install links
This links to the comprehensive docs for installing, instead of arbitrarily choosing different categories to mention in our blog post vs. our docs. (The blog post had "Local/AWS/DigitalOcean/Kubernetes cluster" vs. the docs' "Single-container server/Docker Compose/Kubernetes". The latter is better IMO.)

It also adds a link to upgrade, which was mentioned in the blog post but was not actually linked.