distribution: Weekly team updates
While the document describes updates as a bi-weekly process, we are effectively doing weekly updates. Bi-weekly are too little given that will produce 2 updates on each iteration. The reasons for async and weekly are still valid from the bi-weekly ones, so I kept them.

> - Weekly updates make it easy to forget what you did at the start of the week, forget to ask for help, and make it hard for others to offer help in a timely fashion.

I have removed this point as it was not clear to me "why" weekly updates would cause this.