Remove redlines
A11y has been a good addition, thanks for adding. 

I'm not sold on the redlines plugin yet, I was unsatisfied with it on the CTA project and used a very simple component instead. I want to make sure we are adding things here that have been through the cycle, including through the handoff to dev.