update values
Background: We want our values to be up to date, frequently useful, and frequently referred to. The current list (people, journey, and progress) was well intentioned, and those things still matter, but they were vague and not actually frequently useful or referred to. (We still value people, obviously, but our actual actions, and the more detailed principles in this PR, are what matter, not a word on a page.) In https://github.com/sourcegraph/about/pull/1051 and [RFC 126](https://docs.google.com/document/d/1_L-XOT4oPTOuUkT4_JMykneo_wPn2WqC2BXR6_-U5dI/edit#heading=h.trqab8y0kufp), we reevaluated them and started to come up with more useful values. That effort stalled because it (like our previous values) sought to cover everything that mattered, which was very difficult.

This update to our values starts over from scratch and intends only to cover the values we actually frequently find useful and refer to. It's bootstrapped by values that @nicksnyder and I wrote that we have verbally referred to frequently in recent weeks.

Reviewing this PR means reviewing 2 things:

1. The new definition/scope of values: (new) listing what's useful vs. (old) trying to cover everything that matters.
1. The 3 initial values listed here that we've seeded this page with.

If you have other values you'd like to propose, make a PR on top of this branch or wait until this PR is merged to propose additions.