Add suggested reviewers and update terminology
As our team grows, it's harder for new team members (as well as existing team members) to identify which team members would be able to provide the best feedback for an RFC. This PR adds a new section to the metadata, "Suggested reviewers," which gives readers the chance to suggest specific reviewers to the author. This will make it easier for team members to seek feedback without missing important perspectives.

As well, in our RFC template, we use the term "Requested reviewers" rather than "Required approvers," so this PR also updates that terminology.