Add cs.dev shortcut URL blog post + related styles
Also adds a new blog post template type: `style: short-inline-title` posts, where the title is just shown as a bold first sentence, and there is never a `Read more` link. This is for posts that don't need to be long.

(It's necessary to include the code changes with this PR because Gatsby complains that `style` is not a valid frontmatter property if the code changes are separate from the new content.)