Added an s to http on some images to fix mixed content issue (#1505)
* Updated url in gatsby-config.js file

* Updated gatsby config and node file to fix sitemap errors

Added a trailing slash to permalinks that get generated in gatsby-node.js and changed the canonical-urls plugin option "noTrailingSlash" to false so the canonical links can match the permalinks containing a trailing slash. The reason for the errors on the sitemap are because the permalinks that are getting generated right now in gatsby-node.js are being generated without a trailing slash and is causing a redirect on the browser. For example if you open this link https://about.sourcegraph.com/blog/introducing-code-search-in-sourcegraph and refresh that page a few times you will notice the url starts with a trailing slash then gets removed which is causing a redirect as you can see here https://www.screencast.com/t/0gpAkd5fSNYR These changes should fix the issue. Thanks and let me know if that works!

* Updated gatsby config and node file to fix sitemap errors

* Updated gatsby config and node file to fix sitemap errors

* Added an s to http on some images to fix mixed content issue