update campaigns team goals
Update the goals of the Campaigns team. I marked off what we have done (leaving it ~~strike-through~~ as the context is valuable for now). I also added a line about user testing, as that is an important part of validating that we are ready to open this up to a wider set of customers.