Update engineering-manager-web.md
Hi Nick,

Is the goal to move all our github documentation to the handbook?

See this job description has directions for what we do internally, and that may confuse the candidates. Following edits recommended. 

Changing the language "pick" to "will interview with" as not to confuse candidates that they get to pick who to meet. 

Technical experience github is private. Deleting the link for now and replaced with general description from other job descriptions for now.