Update desired outcome for Distribution goals
We have decided to switch to a different scoring system instead of 1-10. We will use:
- Really Negative
- Negative
- Positive
- Really Positive

The reason for doing this is that its easier to set a desired target at the moment given we have no existing score so setting a target of 9+ could be impossible to achieve and setting a % based increases could be really misleading (If our current score is 1 and we move to a 2, that is a 100% increase but its still a really bad score).