Update code-search-turned-code-checker links
Appending `index:only` to these queries, because it can happen that Go projects in the repo group gets unindexed for whatever reason, and causes things to be slow. Previously, unindexed repositories would just be ignored, but since we support unindexed structural search now, it's best to turn off if there are any in the repogroup.