Add changes for search-core and search-product (#2900)
Our engaging team's docs do not currently contain team-specific contact protocols and are meant more broadly as a filter for who to call on various types of issues. This is an example of what including team-specific contact info here might look like. Additional changes for keywords might be good to edit into this PR