Expand core app norms on getting and giving assistance.
Part of the retro feedback this morning was focused on collaborating
more within the team and eliminating silos of work. We agreed to expand
our team norms within the handbook to ensure that people feel comfortable
knowing when and how to get help, as well as how to make ourselves more
available (e.g. office hours).