Add a playbook for fixing macOS gitserver evictions.
macOS treats space used by Time Machine local snapshots as available
depending on which storage view you use. This leads to a situation where
gitserver is evicting repositories due to disk pressure, but tools like
Finder report that plenty of space is available.

This playbook explains how to diagnose and resolve that situation.