Deprecate and redirect
@rebeccadee do you know what the best way is to delete this post and redirect traffic to the new one? 

maybe @csbailey5t knows? we're looking to combine two posts and redirect any lingering organic traffic from this one (the content of which is to be merged into another post https://github.com/sourcegraph/about/pull/4973 ) to the new one. I figure it has something to do with a 301 redirect but I don't know how to actually implement that 😅